Move mongo classes to own package
