--- src/main/java/com/ericsson/ei/handlers/MongoQuery.java
@@ -1,9 +1,0 @@
-package com.ericsson.ei.handlers;
-
-public interface MongoQuery {
-
-    public String getQueryString();
-
-    @Override
-    public String toString();
-}
