+++ src/main/java/com/ericsson/ei/mongo/MongoQuery.java
@@ -1,39 +1,22 @@
 /*
-   Copyright 2017 Ericsson AB.
+   Copyright 2019 Ericsson AB.
    For a full list of individual contributors, please see the commit history.
-
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at
-
        http://www.apache.org/licenses/LICENSE-2.0
-
    Unless required by applicable law or agreed to in writing, software
    distributed under the License is distributed on an "AS IS" BASIS,
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.ericsson.ei.repository;
+package com.ericsson.ei.mongo;
 
-import java.util.ArrayList;
+public interface MongoQuery {
 
-import org.bson.Document;
-import org.json.JSONException;
+    public String getQueryString();
 
-import com.ericsson.ei.mongo.MongoDBHandler;
-import com.ericsson.ei.mongo.MongoQuery;
-
-public interface ISubscriptionRepository {
-
-    ArrayList<String> getSubscription(MongoQuery query);
-
-    Document modifySubscription(String stringSubscription, String subscriptionName) throws JSONException;
-
-    boolean deleteSubscription(MongoQuery queary);
-
-    void addSubscription(String subscription);
-
-    MongoDBHandler getMongoDbHandler();
-
+    @Override
+    public String toString();
 }
