+++ src/test/java/com/ericsson/ei/handlers/test/StatusHandlerTest.java
@@ -25,9 +25,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import com.ericsson.ei.handlers.MongoDBHandler;
 import com.ericsson.ei.handlers.StatusHandler;
 import com.ericsson.ei.listeners.RMQConnectionListener;
+import com.ericsson.ei.mongo.MongoDBHandler;
 import com.ericsson.ei.status.Status;
 import com.fasterxml.jackson.databind.JsonNode;
