+++ src/test/java/com/ericsson/ei/queryservice/test/QueryServiceTest.java
@@ -38,8 +38,8 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import com.ericsson.ei.App;
-import com.ericsson.ei.handlers.MongoDBHandler;
 import com.ericsson.ei.handlers.ObjectHandler;
+import com.ericsson.ei.mongo.MongoDBHandler;
 import com.ericsson.ei.queryservice.ProcessAggregatedObject;
 import com.ericsson.ei.queryservice.ProcessMissedNotification;
 import com.ericsson.ei.test.utils.TestConfigs;
